+++ backend/src/main/java/com/apex/backend/dto/PaperPositionDTO.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
 
 @Data
 @Builder
@@ -12,8 +13,8 @@ import lombok.NoArgsConstructor;
 public class PaperPositionDTO {
     private String symbol;
     private Integer quantity;
-    private Double entryPrice; // ✅ Renamed from avgPrice to match Controller
-    private Double ltp;        // ✅ Added
-    private Double pnl;        // ✅ Added
-    private Double pnlPercent; // ✅ Added
+    private BigDecimal entryPrice; // ✅ Renamed from avgPrice to match Controller
+    private BigDecimal ltp;        // ✅ Added
+    private BigDecimal pnl;        // ✅ Added
+    private BigDecimal pnlPercent; // ✅ Added
 }
