+++ backend/src/main/java/com/apex/backend/model/UserProfile.java
@@ -6,6 +6,7 @@ import lombok.Builder;
 import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -14,9 +15,9 @@ import java.util.List;
 @AllArgsConstructor
 public class UserProfile {
     private String name;
-    private double availableFunds;
-    private double totalInvested;
-    private double currentValue;
-    private double todaysPnl;
+    private BigDecimal availableFunds;
+    private BigDecimal totalInvested;
+    private BigDecimal currentValue;
+    private BigDecimal todaysPnl;
     private List<Object> holdings;
 }
